# src/main/java/com/caseplan/adapter/in/web/dto/CreateCasePlanRequest.java
package com.caseplan.adapter.in.web.dto;

import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.validation.constraints.NotBlank;

@Getter
@Setter
@NoArgsConstructor
public class CreateCasePlanRequest {
    @NotBlank(message = "Client first name is required")
    private String clientFirstName;

    @NotBlank(message = "Client last name is required")
    private String clientLastName;

    private String clientIdNumber;

    @NotBlank(message = "Attorney name is required")
    private String attorneyName;

    @NotBlank(message = "Bar number is required")
    private String barNumber;

    private String referringSource;

    // Optional at intake stage; many new cases do not have an assigned case number yet.
    private String caseNumber;

    @NotBlank(message = "Primary cause of action is required")
    private String primaryCauseOfAction;

    private String opposingParty;

    @NotBlank(message = "Legal remedy sought is required")
    private String remedySought;

    private String additionalCauses;
    private String priorLegalActions;
    private String caseDocuments;
    private Boolean confirm;
}