# src/main/java/com/caseplan/adapter/in/web/dto/UpdateClientRequest.java
package com.caseplan.adapter.in.web.dto;

import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.validation.constraints.NotBlank;

@Getter
@Setter
@NoArgsConstructor
public class UpdateClientRequest {

    @NotBlank(message = "Client first name is required")
    private String firstName;

    @NotBlank(message = "Client last name is required")
    private String lastName;

    private String idNumber;
}